+++ src/main/java/ch/epfl/leb/sass/models/photophysics/FluorophoreDynamicsBuilder.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package ch.epfl.leb.sass.models.fluorophores.internal.dynamics;
+package ch.epfl.leb.sass.models.photophysics;
 
 /**
  * Interface for creating fluorophore dynamics.